importFromPanorama(): fix harmless int32 overflow
git-svn-id: https://svn.osgeo.org/gdal/trunk@32650 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogr_srs_panorama.cpp
@@ -343,7 +343,7 @@ OGRErr OGRSpatialReference::importFromPanorama( long iProjSys, long iDatum,
                 else
                 {
                     nZone = (int) padfPrjParams[7];
-                    dfCenterLong = 6 * nZone - 3;
+                    dfCenterLong = 6 * (double)nZone - 3;
                 }
 
                 padfPrjParams[5] = nZone * 1000000.0 + 500000.0;
